give Primitive* its on file
